+++ src/main/java/hackatona/model/User.java
@@ -20,7 +20,7 @@ public class User implements Serializable, UserDetails {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id")
-	private Long id;
+	private Integer id;
 
 	private String nome;
 
@@ -30,11 +30,11 @@ public class User implements Serializable, UserDetails {
 
 	private Integer perfil;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
